feat: users apo 및 type 구현
