+++ app/_types/user.type.ts
@@ -1,0 +1,35 @@
+interface CreateUserParams {
+  email: string;
+  nickname: string;
+  password: string;
+}
+
+interface UserType {
+  id: number;
+  email: string;
+  nickname: string;
+  profileImageUrl: string;
+  createdAt: string;
+  updatedAt: string;
+}
+
+interface UpdateUserParams {
+  nickname: string;
+  profileImageUrl: string;
+}
+
+interface CreateProfileImageParams {
+  image: File;
+}
+
+interface ProfileImageType {
+  profileImageUrl: string;
+}
+
+export type {
+  UserType,
+  ProfileImageType,
+  CreateUserParams,
+  UpdateUserParams,
+  CreateProfileImageParams,
+};
